[Fix] 위치 권한 설정 수정
